First pass code reorg
